requirements.txt
pydantic
PyGithub